Update linkAction for template rendering
